Move Windows config recipes into new directory
This makes it easy to test those recipes using local sandboxes, so there
won't be any surprises when attempting to generate an official build.

This also cleans up the 32-bit JRE detection on Windows.  CMake doesn't
differentiate between 32-bit and 64-bit JRE's, but we have to use a
32-bit JRE to run the unit tests for a 32-bit build, so the easiest way
to do that is to create a symlink to the 32-bit JRE in a known location
(as we do on Linux) and look for the JRE at that location.