Bitcode on iOS, remove ARMv6 & Xcode 5.1.x req
The purpose of this commit is three-fold:

(1) Remove support for ARMv6 from our official binaries.  The last
iDevice to require ARMv6 was the iPhone 3G, which required iOS 4.2.1 or
older.  Our binaries have always required iOS 4.3 or newer, so I'm not
sure if the ARMv6 fork of our binaries was ever useful to begin with.
In any case, if it ever was useful, it no longer is.  This removes the
Xcode 4.3.x requirement.

(2) Embed bitcode into the iOS binaries, which facilitates building
applications for iOS 9 and later (bitcode is required for watchOS and
tvOS apps and is the default when building iOS apps in Xcode 7 and
later.)  This necessitates building the iOS binaries with Xcode 7.2.x,
which uses Clang.  Although Clang doesn't do a very good job of
optimizing some of libjpeg-turbo's algorithms, the addition of
NEON-accelerated Huffman encoding in libjpeg-turbo 1.5.0 means that we
can switch to Clang without too much of a performance penalty (refer to
libjpeg-turbo/libjpeg-turbo#5).  The use of bitcode means that the
libjpeg-turbo binaries can no longer be used to build apps for
iOS < 6.0, but any apps under active development will have long since
had to abandon those old iOS releases anyhow.

(3) Remove unnecessary Xcode 5.1.x requirement.  Building with GCC 5 and
-mmacosx-version-min=10.5 is sufficient to provide backward
compatibility with Leopard, since libjpeg-turbo has no external
dependencies other than libgcc and libSystem.

Fixes libjpeg-turbo/libjpeg-turbo#5.