Build x86-64 OS X/macOS binaries using Xcode
+ increase the minimum supported OS X version for the x86-64 binaries to
10.7.

GCC's implementation of thread-local storage isn't backward compatible
with Xcode 7.2.x and prior, which effectively meant that Xcode couldn't
be used with the official libjpeg-turbo 2.0.5 binaries except on OS X
10.11 "El Capitan" or later.  Enabling thread-local storage with Xcode
requires targetting OS X 10.7 "Lion" or later, but this should not be an
issue, since all 64-bit Intel Macs are compatible with Lion.

With the x86-64 libjpeg-turbo code, Xcode 7.2 is performance-neutral
relative to GCC 5.