buildljt -v: Fail if subscript fails
If -v was passed to buildljt (which our CI system always does), then the
exit status of buildljt was the exit status of tee, not the exit status
of buildljt.$PLATFORM.  Effectively, that meant that errors in the
official build jobs were not communicated to GitHub Actions.
'set -o pipefail' causes the exit status of buildljt to be the logical
AND of the exit statuses of buildljt.$PLATFORM and tee.