Mac: Include Java wrapper in secondary build
I'm honestly not sure why the TurboJPEG Java wrapper was disabled in the
secondary build, but that has been the case since the Armv8 architecture
was added to the official macOS binaries in
02801e10c0ac87810a8c0cb19854b40fd798d756 (libjpeg-turbo 2.1 beta1.)  I
can only guess that it was due to limitations in the build system that
was used at the time.  That behavior was carried over into
18112a28383af976d2b2495212f96a316cdde960.  Thus, when using an
official build of libjpeg-turbo 2.1.x-3.1 beta1 (which were built on an
Intel Mac), or when using a CI build, the TurboJPEG Java wrapper could
not be used with the Arm64 architecture.  When using an official build
of libjpeg-turbo 3.1.0 and later (which were built on an Apple Silicon
Mac), the TurboJPEG Java wrapper could not be used with the x86-64
architecture.